\name{rQESN-package}
\alias{rQESN-package}
\alias{rQESN}
\docType{package}
\title{
\packageTitle{rQESN}
}
\description{
\packageDescription{rQESN}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rQESN}
\packageIndices{rQESN}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rQESN}

Maintainer: \packageMaintainer{rQESN}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
